% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPC.R
\name{CalculateZoneBorders}
\alias{CalculateZoneBorders}
\title{CalculateZoneBorders}
\usage{
CalculateZoneBorders(limits, controlLimitDistance = 3)
}
\arguments{
\item{limits}{List of limits provided by \code{\link{CalculateLimits}}}

\item{controlLimitDistance}{Multiple of st.dev to be used to calculate limits, possible values: 1, 2, 3 (default); this parameter affect the interpretation of lcl and ucl parameters}
}
\value{
Vector of zones
}
\description{
Some Nelson rules uses so-called zones. This function calculates the borders of the zones for given limits.
}
\examples{
limits = CalculateLimits(x = rnorm(10), lcl = NA, cl = 100, ucl = NA, type = 'i')
CalculateZoneBorders(limits)
#limits is object created by CalculateLimits() function
}
