% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chart.R
\name{make_chart}
\alias{make_chart}
\title{Create a Customized SPC Chart}
\usage{
make_chart(
  data,
  chart_title = "",
  chart_title_size = 14,
  caption = "",
  caption_size = 8
)
}
\arguments{
\item{data}{A data frame containing the SPC data including at least the columns 'x' for dates,
'y' for the measurement values, 'cl' for the centerline, 'lcl' for the lower control limit,
and 'ucl' for the upper control limit.}

\item{chart_title}{A character string representing the title of the chart.}

\item{chart_title_size}{Numeric value defaulted at 14 but can be changed according to need}

\item{caption}{Character string that can be used to enter source of the data on bottom right}

\item{caption_size}{Numeric value that is defaulted at 8 but can be used to change size of caption}
}
\value{
A ggplot object representing the SPC chart with categorical date handling and
customized aesthetics.
}
\description{
This function generates a statistical process control (SPC) chart from provided SPC data.
It ensures that date columns are treated as categorical variables, allowing each unique
date to be plotted without summarization. The function colors various components of the
chart and uses discrete scaling for the x-axis to display all unique dates.
}
\examples{
data <- data.frame(
  x = seq(as.Date("2021-01-01"), by = "month", length.out = 12),
  y = rnorm(12, 100, 10),
  cl = 100,
  lcl = 85,
  ucl = 115
)
chart <- make_chart(data, "Monthly SPC Chart", 15, "Source: Imaginary database", 10)
print(chart)
}
