% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPC.R
\name{NelsonRules}
\alias{NelsonRules}
\title{NelsonRules}
\usage{
NelsonRules(ruleN, data, zoneB, limits, parRules = NULL, ...)
}
\arguments{
\item{ruleN}{Name of individual Rule function "Rule1" to "Rule8"}

\item{data}{Data to be checked, numerical vector}

\item{zoneB}{Vector of zones created by \code{\link{CalculateLimits}}}

\item{limits}{List of limit created by \code{\link{CalculateLimits}}}

\item{parRules}{List of optional parameters for this particular rule}

\item{...}{unspecified arguments of a function}
}
\value{
Result of individual Rule function with predefined parameters
}
\description{
Auxiliary function to calling individual Rule functions.
}
\details{
Handling the missing values:

Missing values are represented by the symbol NA - not available.

Rule 1:
 NAs do not violate this rule.

Rule 2-8:
  NAs are ignored, they do not break Rule evaluation.
  NA values are removed from the vector, the rule function is calculated
  and then the NAs are returned back to it's original position in the vector.
}
