% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spc_chart.R
\name{plot_spc_chart}
\alias{plot_spc_chart}
\title{Create a Customized SPC Chart with Phase Handling and Annotations}
\usage{
plot_spc_chart(
  data,
  chart_title = "",
  chart_title_size = 14,
  caption = "",
  caption_size = 8,
  annotations = NULL
)
}
\arguments{
\item{data}{A data frame containing the SPC data including at least the columns 'x' for dates,
'y' for the measurement values, 'cl' for the centerline, 'lcl' for the lower control limit,
'ucl' for the upper control limit, and optionally 'phase' for indicating phase changes.}

\item{chart_title}{A character string representing the title of the chart.}

\item{chart_title_size}{Numeric value defaulted at 14 but can be changed according to need.}

\item{caption}{Character string that can be used to enter the source of the data on the bottom right.}

\item{caption_size}{Numeric value that is defaulted at 8 but can be used to change the size of the caption.}

\item{annotations}{A data frame containing the annotations with columns 'row_number', 'label',
'text_size', 'position_x', 'position_y'. The 'position_x' and 'position_y' columns
specify the offset for the annotation text relative to the point.}
}
\value{
A ggplot object representing the SPC chart with categorical date handling and
customized aesthetics, including phase handling.
}
\description{
This function generates a statistical process control (SPC) chart from provided SPC data.
It ensures that date columns are treated as categorical variables, allowing each unique
date to be plotted without summarization. The function colors various components of the
chart and uses discrete scaling for the x-axis to display all unique dates. If phase changes
exist, it discontinues line connections between different phases. Additionally, it allows
for custom annotations on the chart.
}
\examples{
data <- data.frame(
  x = seq(as.Date("2021-01-01"), by = "month", length.out = 12),
  y = rnorm(12, 100, 10),
  cl = 100,
  lcl = 85,
  ucl = 115,
  phase = rep(1:2, each = 6)
)
annotations <- data.frame(
  row_number = c(3, 9),
  label = c("Annotation 1", "Annotation 2"),
  text_size = c(4, 4),
  position_x = c(0.2, 0),
  position_y = c(10, 10)
)
chart <- plot_spc_chart(data, "Monthly SPC Chart", 15, "Source: Imaginary database", 10, annotations)
print(chart)
}
