% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPC.R
\name{Rule1}
\alias{Rule1}
\title{Rule 1}
\usage{
Rule1(x, lcl, ucl, sides, ...)
}
\arguments{
\item{x}{Numerical vector}

\item{lcl}{Lower control limit, single number}

\item{ucl}{Upper control limit, single number}

\item{sides}{Monitored side of the process: either "two-sided" (default), "upper" or "lower"}

\item{...}{unspecified arguments of a function}
}
\value{
Vector of the same length as x
}
\description{
One point beyond the control limits
}
\details{
0 means: ok \cr
1 means: violation \cr

inequality used during evaluation

parametr sides is internally encoded as: 1 for "two-sided", 2 for "upper", 3 for "lower"
}
\examples{
Rule1(x = rnorm(10), lcl = 10, ucl = 100, sides = "two-sided")
}
