% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPC.R
\name{Rule2}
\alias{Rule2}
\title{Rule 2}
\usage{
Rule2(x, cl, nPoints = 9, ...)
}
\arguments{
\item{x}{Numerical vector}

\item{cl}{central line, single number}

\item{nPoints}{Sequence of consequtive points to be evaluated}

\item{...}{unspecified arguments of a function}
}
\value{
Vector of the same length as x
}
\description{
Nine points in a row are on one side of the central line.
}
\details{
0 means: ok \cr
1 means: violation \cr

inequality used during evaluation
}
\examples{
Rule2(x = rnorm(20), cl=0, nPoints = 9)
}
