% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPC.R
\name{SetParameters}
\alias{SetParameters}
\title{SetParameters}
\usage{
SetParameters()
}
\value{
List of optional parameters
}
\description{
Creates optional parameters with default settings.
}
\details{
The function is called without any parameter. If you want to modify any or the rules' setting, modify the result of this function and plug it to \code{\link{EvaluateRules}}'s parRules parameter.
}
\examples{
pars <- SetParameters()
pars$Rule1$sides <- "upper"
#function doos not need any input parameters
}
