% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructured_data_qic.R
\name{create_spc_data}
\alias{create_spc_data}
\title{Create SPC Data}
\usage{
create_spc_data(data, date_col, value_col, chart_type)
}
\arguments{
\item{data}{Data frame containing the data}

\item{date_col}{Name of the date column}

\item{value_col}{Name of the value column}

\item{chart_type}{Type of SPC chart to create}
}
\value{
Returns a data frame with SPC data including adjusted LCL values.
}
\description{
This function creates a Statistical Process Control chart data and ensures LCL values are not negative.
}
\examples{
data <- data.frame(date = seq(as.Date('2020-01-01'), by = 'month', length.out = 30),
                   value = rnorm(30, 100, 15))
create_spc_data(data, 'date', 'value', 'xbar')
}
