% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret}
\alias{interpret}
\title{Interpret SPC Data}
\usage{
interpret(data)
}
\arguments{
\item{data}{A data frame containing the SPC data including at least the columns 'x' for dates,
'y' for the measurement values, and additional calculated columns for analysis.}
}
\value{
A data frame with columns "Special Cause Variation" and "Duration" indicating the type
of variation and the corresponding range of dates.
}
\description{
This function interprets the SPC data to detect special cause variations such as shifts, trends,
15+ consecutive points, and sigma signals. It outputs a table summarizing these findings with
the range of occurrences.
}
