% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_control_chart.R
\name{make_chart}
\alias{make_chart}
\title{Create an SPC Chart with Custom Aesthetics}
\usage{
make_chart(data, chart_title)
}
\arguments{
\item{data}{A data frame containing the columns x, y, cl, lcl, and ucl.}

\item{chart_title}{Title of the chart to be displayed at the top.}
}
\value{
A ggplot object representing the SPC chart.
}
\description{
This function generates a customized Statistical Process Control (SPC) chart using ggplot2.
The chart includes the main data line, control limits (CL, UCL, LCL) and points, with specific annotations
and a defined color palette. It returns a ggplot object ready for displaying or further customization.
}
\examples{
data <- data.frame(
  x = 1:30,
  y = rnorm(30, 100, 15),
  cl = rep(100, 30),
  ucl = rep(115, 30),
  lcl = rep(85, 30)
)
make_chart(data, "Sample SPC Chart")
}
