% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_test.R
\name{run_test}
\alias{run_test}
\title{Run SPC Chart Test with Advanced Highlighting and Phase Handling}
\usage{
run_test(
  data,
  chart_title = "",
  chart_title_size = 14,
  caption = "",
  caption_size = 8
)
}
\arguments{
\item{data}{A data frame containing the SPC data including at least the columns 'x' for dates,
'y' for the measurement values, 'cl' for the centerline, 'lcl' for the lower control limit,
'ucl' for the upper control limit, and various signal columns.}

\item{chart_title}{A character string representing the title of the chart.}

\item{chart_title_size}{Numeric value defaulted at 14 but can be changed according to need}

\item{caption}{Character string that can be used to enter source of the data on bottom right}

\item{caption_size}{Numeric value that is defaulted at 8 but can be used to change size of caption}
}
\value{
A ggplot object representing the SPC chart with customized aesthetics.
}
\description{
This function produces an SPC chart from the data frame generated by \code{create_spc_data}, using the styling defined in \code{make_chart}.
It highlights points based on various conditions: \code{sigma.signal}, \code{shift} sequences, \code{fifteen_more} long runs, \code{trend} stability, and \code{two_more} patterns.
If phases are present, it treats each phase as an independent subgroup for applying these highlighting rules.
}
\examples{
data <- create_spc_data(your_data_frame, 'date', 'value', 'xbar')
run_test(data, "Comprehensive Monthly SPC Chart")
}
